refac: `get_wholesale_metering_point_times_series` as module
