+++ source/databricks/calculation_engine/tests/calculation/energy/test_transform_hour_to_quarter.py
@@ -11,59 +11,25 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
+from decimal import Decimal
 
 import pytest
-from decimal import Decimal
-from datetime import datetime
 from pyspark.sql import SparkSession, Row
-from pyspark.sql.types import (
-    TimestampType,
-)
 
-from package.constants import Colname
-from package.codelists import MeteringPointResolution, QuantityQuality
+from calculation.basis_data_time_series_points_factories import (
+    basis_data_time_series_points_row,
+)
 from package.calculation.energy.hour_to_quarter import (
     transform_hour_to_quarter,
     metering_point_time_series_schema,
 )
-
-
-def basis_data_time_series_points_row(
-    grid_area: str = "805",
-    to_grid_area: str = "805",
-    from_grid_area: str = "806",
-    metering_point_id: str = "the_metering_point_id",
-    metering_point_type: str = "the_metering_point_type",
-    resolution: MeteringPointResolution = MeteringPointResolution.HOUR,
-    observation_time: TimestampType() = datetime(2020, 1, 1, 0, 0),
-    quantity: Decimal = Decimal("4.444444"),
-    quality: QuantityQuality = QuantityQuality.ESTIMATED,
-    energy_supplier_id: str = "the_energy_supplier_id",
-    balance_responsible_id: str = "the_balance_responsible_id",
-    settlement_method: str = "the_settlement_method",
-) -> Row:
-    row = {
-        Colname.grid_area: grid_area,
-        Colname.to_grid_area: to_grid_area,
-        Colname.from_grid_area: from_grid_area,
-        Colname.metering_point_id: metering_point_id,
-        Colname.metering_point_type: metering_point_type,
-        Colname.resolution: resolution.value,
-        Colname.observation_time: observation_time,
-        Colname.quantity: quantity,
-        Colname.quality: quality.value,
-        Colname.energy_supplier_id: energy_supplier_id,
-        Colname.balance_responsible_id: balance_responsible_id,
-        Colname.settlement_method: settlement_method,
-    }
-
-    return Row(**row)
+from package.codelists import MeteringPointResolution
+from package.constants import Colname
 
 
 def test__transform_hour_to_quarter__when_invalid_input_schema__raise_assertion_error(
     spark: SparkSession,
-):
+) -> None:
     # Arrange
     basis_data_time_series_points = spark.createDataFrame(
         data=[Row(**({"Hello": "World"}))]
